Move StartScheduler command to tasks.py
